+++ misc/python/python-lambda/frinx_worker/python_lambda/__init__.py
@@ -16,7 +16,6 @@ from frinx_worker.python_lambda.code_wrapper import executed_func_template
 
 
 class PythonLambda(WorkerImpl):
-
     VISITOR = FunctionCallVisitor()
 
     class ExecutionProperties(TaskExecutionProperties):
@@ -25,9 +24,9 @@ class PythonLambda(WorkerImpl):
         pass_worker_input_exception_to_task_output: bool = True
 
     class WorkerDefinition(TaskDefinition):
-        name: str = 'PYTHON_lambda'
-        description: str = 'Execute simple python code'
-        labels: list[str] = ['PYTHON']
+        name: str = "PYTHON_lambda"
+        description: str = "Execute simple python code"
+        labels: list[str] = ["PYTHON"]
         timeout_seconds: int = 60
         response_timeout_seconds: int = 59
 
@@ -40,9 +39,7 @@ class PythonLambda(WorkerImpl):
 
     @staticmethod
     def _wrap_lambda_func(worker_input: WorkerInput) -> str:
-        return executed_func_template.format(
-            custom_execution_commands=worker_input.lambda_function
-        )
+        return executed_func_template.format(custom_execution_commands=worker_input.lambda_function)
 
     @classmethod
     def _validate_code(cls, code: str) -> None:
@@ -51,18 +48,13 @@ class PythonLambda(WorkerImpl):
     @staticmethod
     @timeout(5)  # type: ignore[misc]
     def _process_code(code: str, worker_inputs: DictAny) -> DictAny:
-        ex_locals: dict[str, Any] = {'worker_input': worker_inputs}
+        ex_locals: dict[str, Any] = {"worker_input": worker_inputs}
         exec(code, None, ex_locals)
-        return ex_locals.get('result', {})
+        return ex_locals.get("result", {})
 
     def execute(self, worker_input: WorkerInput) -> TaskResult[Any]:
         code = self._wrap_lambda_func(worker_input)
         self._validate_code(code)
 
         result = self._process_code(code, worker_input.worker_inputs)
-        return TaskResult(
-            status=TaskResultStatus.COMPLETED,
-            output=self.WorkerOutput(
-                result=result
-            )
-        )
+        return TaskResult(status=TaskResultStatus.COMPLETED, output=self.WorkerOutput(result=result))